Se modifica interfaz del login. Se corrigen errores.
